pyproject.toml
[tool.poetry]
name = "py-todo"
version = "0.1.0"
description = "To Do backend service using Python/FastAPI"
authors = ["Mohsen <mim.khodabakhshi@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.21.4"}
pydantic-settings = "2.1.0"
pyjwt = "^2.8.0"
aerich = "^0.7.2"
pydantic = {extras = ["email"], version = "^2.8.2"}
cryptography = "^42.0.8"



[tool.aerich]
tortoise_orm = "services.database.connection.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
